对4G以上文件的支持

+++ File-Encryption-AES/File-Encryption-AES/FileFunction.h
@@ -6,7 +6,7 @@
 using namespace std;
 
 
-long SizeOfFile(char * file_loc);
+long long SizeOfFile(char * file_loc);
 void ReadBuffFromFile(unsigned char * buff, fstream &infile);
 void WriteDataFromBuff(unsigned char * buff, ofstream &ofile);
 void ClearBuff(char * Buff);
